feat: refactoring read_env_from to module kubernetes_services
