+++ app/v1/modules/kubernetes_services/__init__.py
@@ -1,11 +1,13 @@
 # -*- coding: utf-8 -*-
 
 
+from .script_read_env_from import ScriptEnvironReaderFromService
 from .script_to_dict import ScriptToDictParserService
 from .script_read_env import ScriptEnvironReaderService
 
 
 __all__ = [
+    "ScriptEnvironReaderFromService",
     "ScriptToDictParserService",
     "ScriptEnvironReaderService",
 ]
